+++ src/config/modelProviders/index.ts
@@ -8,6 +8,7 @@ import MistralProvider from './mistral';
 import MoonshotProvider from './moonshot';
 import OllamaProvider from './ollama';
 import OpenAIProvider from './openai';
+import OpenRouterProvider from './openrouter';
 import PerplexityProvider from './perplexity';
 import ZhiPuProvider from './zhipu';
 
@@ -20,6 +21,7 @@ export const LOBE_DEFAULT_MODEL_LIST: ChatModelCard[] = [
   MistralProvider.chatModels,
   MoonshotProvider.chatModels,
   OllamaProvider.chatModels,
+  OpenRouterProvider.chatModels,
   PerplexityProvider.chatModels,
   AnthropicProvider.chatModels,
 ].flat();
@@ -32,5 +34,6 @@ export { default as MistralProvider } from './mistral';
 export { default as MoonshotProvider } from './moonshot';
 export { default as OllamaProvider } from './ollama';
 export { default as OpenAIProvider } from './openai';
+export { default as OpenRouterProvider } from './openrouter';
 export { default as PerplexityProvider } from './perplexity';
 export { default as ZhiPuProvider } from './zhipu';